+++ frontend/src/components/Notification/Notification.styled.ts
