+++ frontend/src/components/TeamCalendar/TeamCalendar.styled.ts
