+++ frontend/src/components/TeamPlaceMenu/TeamPlaceMenu.styled.ts
