+++ frontend/src/components/Thread/Thread.styled.ts
